refactor: extract AI message streaming logic to separate service
- Create aiMessageService.ts with dedicated functions for AI message handling
- Split createAndStreamAIMessage into 4 focused helper functions to reduce cognitive complexity
- Move API call, streaming, placeholder creation, and error handling to service layer
- Simplify Chat.tsx by removing 150+ lines of complex streaming logic
- Improve code maintainability and testability through better separation of concerns
