+++ frontend/apps/app/components/Chat/services/index.ts
@@ -1,1 +1,2 @@
 export * from './messageServiceClient'
+export * from './aiMessageService'
